Fix "DeniedRequiresTLS" value for activation request
The value corresponding to a Negative Response "Requires TLS" after sending a Routing Activation Request is invalid.

Currently this value is set to "0x06", but according to the ISO-13400 standard, the "0x06" value corresponds to an unknown activation type.

Referring to the standard, we can state that the Negative Response Code "Requires TLS" should be set to "0x07" :

> After reception of a routing activation request message with a routing activation type, which requires the
> secure TLS connection to the DoIP entity then each DoIP entity supporting the secure TCP communication
> shall send the routing activation response message with the response code set to 0x07.

This patch applies the correct correction to the "DeniedRequiresTLS" value.